Models CRUD added

+++ src/main/kotlin/tech/developer/catalogodecarros/service/MakeServiceImpl.kt
@@ -3,21 +3,25 @@ package tech.developer.catalogodecarros.service
 import org.springframework.stereotype.Service
 import tech.developer.catalogodecarros.dto.MakeDTO
 import tech.developer.catalogodecarros.repository.MakeRepository
+import tech.developer.catalogodecarros.repository.ModelRepository
 import tech.developer.catalogodecarros.utils.exceptions.MakeException
 import tech.developer.catalogodecarros.utils.mapper.MakeMapper
 
 @Service
 class MakeServiceImpl(
     private val makeRepository: MakeRepository,
+    private val modelRepository: ModelRepository,
     private var makeMapper: MakeMapper
 ) : MakeService {
     @Suppress("DEPRECATED_IDENTITY_EQUALS")
     override fun createMake(makeDTO: MakeDTO): MakeDTO {
 
-        if (makeDTO.id !== -1) throw IllegalArgumentException("ID is not required in the request")
+        if (makeDTO.id != -1) throw IllegalArgumentException("ID is not required in the request")
+
+        if (makeDTO.name.isEmpty()) throw IllegalArgumentException("Name cannot be empty")
 
         val make = makeRepository.save(makeMapper.toEntity(makeDTO))
-        return makeMapper.fromEntity(make);
+        return makeMapper.fromEntity(make)
     }
 
     override fun getMakes(): List<MakeDTO> {
@@ -50,9 +54,12 @@ class MakeServiceImpl(
 
     override fun deleteMake(id: Int) {
         val exists = makeRepository.existsById(id)
-
         if (!exists)
             throw MakeException("Make with id $id is not present")
+
+        val existsInModel = modelRepository.existsByMakeId(id)
+        if (existsInModel)
+            throw MakeException("Make with id $id is present in models list")
 
         makeRepository.deleteById(id)
     }
